+++ qim3d/viz/volume_explorer/__init__.py
@@ -1,1 +1,3 @@
-from .run import volume_explorer
+from .run import volume_explorer, itk_vtk
+
+__all__ = ['volume_explorer']