+++ qim3d/viz/volume_explorer/helpers.py
@@ -17,29 +17,23 @@ WINDOWS = 'Windows'
 MAC = 'Darwin'
 
 
-def get_itk_dir() -> Path:
-    qim_dir = Path(qim3d.__file__).parents[0]  # points to .../qim3d/qim3d/
-    dir = qim_dir.joinpath('viz/volume_explorer')
-    return dir
+def get_volume_explorer_dir() -> Path:
+    """Return the path to the bundled Volume Explorer assets inside qim3d."""
+    qim_dir = Path(qim3d.__file__).parents[0]
+    return qim_dir.joinpath('viz/volume_explorer')
 
 
-def get_nvm_dir(dir: Path = None) -> Path:
-    if platform.system() in [LINUX, MAC]:
-        following_folder = '.nvm'
-    elif platform.system() == WINDOWS:
-        following_folder = ''
-    return (
-        dir.joinpath(following_folder)
-        if dir is not None
-        else get_qim_dir().joinpath(following_folder)
-    )
+def get_nvm_dir(dir: Path | None = None) -> Path:
+    base_dir = dir or get_volume_explorer_dir()
+    following_folder = '.nvm' if platform.system() in [LINUX, MAC] else ''
+    return base_dir.joinpath(following_folder)
 
 
-def get_node_binaries_dir(nvm_dir: Path = None) -> Path:
+def get_node_binaries_dir(nvm_dir: Path | None = None) -> Path:
     """
     Versions could change in time. This makes sure we use the newest one.
 
-    For windows we have to pass the argument nvm_dir and it is the itk-vtk_dir
+    For Windows we have to pass the argument nvm_dir and it is the volume-explorer dir
     """
     if platform.system() in [LINUX, MAC]:
         following_folder = 'versions/node'
@@ -48,38 +42,28 @@ def get_node_binaries_dir(nvm_dir: Path = None) -> Path:
         following_folder = 'node-versions'
         binaries_folder = 'installation'
 
-    node_folder = (
-        nvm_dir.joinpath(following_folder)
-        if nvm_dir is not None
-        else get_nvm_dir().joinpath(following_folder)
-    )
+    node_folder = (nvm_dir or get_nvm_dir()).joinpath(following_folder)
 
-    # We don't wanna throw an error
+    # We don't want to throw an error
     # Instead we return None and check the returned value in run.py
     if not os.path.isdir(node_folder):
         return None
-    l = sorted(os.listdir(node_folder))
 
-    for name in l[::-1]:
+    for name in sorted(os.listdir(node_folder))[::-1]:
         path = node_folder.joinpath(name)
         if os.path.isdir(path):
             return path.joinpath(binaries_folder)
 
 
-def get_viewer_dir(dir: Path = None) -> Path:
-    following_folder = 'viewer_app'
-    return (
-        dir.joinpath(following_folder)
-        if dir is not None
-        else get_qim_dir().joinpath(following_folder)
-    )
+def get_viewer_dir(dir: Path | None = None) -> Path:
+    base_dir = dir or get_volume_explorer_dir()
+    return base_dir.joinpath('viewer_app')
 
 
-def get_viewer_binaries(viewer_dir: Path = None) -> Path:
+def get_viewer_binaries(viewer_dir: Path | None = None) -> Path:
     following_folder1 = 'node_modules'
     following_folder2 = '.bin'
-    if viewer_dir is None:
-        viewer_dir = get_viewer_dir()
+    viewer_dir = viewer_dir or get_viewer_dir()
     return viewer_dir.joinpath(following_folder1).joinpath(following_folder2)