+++ qim3d/viz/volume_explorer/installation.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from .helpers import (
     SOURCE_FNM,
     NotInstalledError,
-    get_itk_dir,
+    get_volume_explorer_dir,
     get_node_binaries_dir,
     get_nvm_dir,
     get_viewer_dir,
@@ -29,7 +29,7 @@ class Installer:
             self.install_viewer,
         )
 
-        self.dir = get_itk_dir()  # volume_explorer folder within qim3d.viz
+        self.dir = get_volume_explorer_dir()  # volume_explorer folder within qim3d.viz
 
         # If nvm was already installed, there should be this environment variable
         # However it could have also been installed via our process, or user deleted the folder but didn't adjusted the bashrc, that's why we check again
@@ -206,4 +206,4 @@ class Installer:
 
         print('Installing volume-explorer...')
         run_for_platform(linux_func=_linux, windows_func=_windows, macos_func=_linux)
-        print('volume-explorer installed')
+        print('Volume Explorer installed')